feat(core,stage): migrate to molrs 0.14 — identity columns are u64 (Idx)
molrs 0.14 makes the identity/schema-uint column family `Idx = u64`. The JS
names are unchanged, but `setColU32` / `copyColU32` / `viewColU32` / `getU32`
now take and return `BigUint64Array`, and `Block.dtype` reports `"u64"` for
those columns. That is the whole breaking surface — no molrs class or function
was renamed or reshaped, and not one import in this diff changed.

`core/src/domain_uint.ts` is the boundary. GPU meshes, CSR tables, and scene
row maps stay `number` / `Uint32Array`; molrs identity columns are converted at
the edge and nowhere else. `toRowIndex` range-checks rather than silently
wrapping, because an id past `Number.MAX_SAFE_INTEGER` is a contract failure,
not a number to round.

Reaches every consumer of those columns: bond endpoints through the artist and
CSR builders, id alignment in source composition, eight modifiers' dtype
guards, the analysis snapshot/worker protocol, the trajectory worker's stream
reader, the RPC wire (new `"u64"` carrier, legacy `"u32"` decode promoted via
`toDomainUint`), optimize staging, sketch's bond block, and ~45 test fixtures.

molrs 0.14 is not published yet, so both manifests point at the local build.
This is deliberate and temporary — CI `npm ci` cannot resolve a path outside
the repo, and nothing in `check:versions` or `check:molrs-gateway` guards the
molrs pin, so it has to be reverted to `^0.14.0` by hand at publish time.

Known gaps, tracked rather than hidden:
- `python/src/molvis/wire.py` has no `"u64"` in `WireDType` / `_NUMPY_BY_DTYPE`,
  so TS→Python frame decode raises. Python→TS still works via the legacy u32
  branch. `test_wire_parity` fails on this.
- `Block.keys()` ordering changed; `stage/tests/mode/hit_info.test.ts` asserts
  the old order, which molrs never guaranteed.
- The storage-width `"u32"` dtype is reachable (molrs' Zarr reader emits it,
  and molvis loads Zarr) but only half-handled: `SelectTypeModifier` routes it
  into the u64-only `viewColU32`, and `ColorByPropertyModifier` /
  `source_composition` accept it with no read branch.

+++ stage/src/atom_type.ts
@@ -13,7 +13,7 @@ export interface AtomTypeSource {
   hasU32(name: string): boolean;
   dtype(name: string): string | undefined;
   getStr(name: string): string[];
-  getU32(name: string): Uint32Array;
+  getU32(name: string): BigUint64Array;
 }
 
 /**
@@ -34,7 +34,7 @@ export function readAtomTypeKeys(atoms: AtomTypeSource): string[] | undefined {
   }
   const idDtype = atoms.dtype(TYPE_ID);
   if (idDtype !== undefined) {
-    throw new Error(`column '${TYPE_ID}' must be u32, got '${idDtype}'`);
+    throw new Error(`column '${TYPE_ID}' must be u64, got '${idDtype}'`);
   }
   return undefined;
 }
